requirements.txt
google-cloud-firestore
google-cloud-aiplatform
functions-framework
